tests/cases/tree-construction/mensbeam03.dat

#data
<!DOCTYPE html><body></body> te st
#errors
(1,30): unexpected-char-after-body
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     " te st"

#data
X
#errors
#document-fragment
template
#document
| "X"
